refactor(basectl): extract TimestampJson into basectl-cli output module
