+++ crates/infra/basectl/src/lib.rs
@@ -15,10 +15,10 @@ mod output;
 pub use output::{
     COLOR_ACTIVE_BORDER, COLOR_BASE_BLUE, COLOR_BURN, COLOR_GAS_FILL, COLOR_GROWTH,
     COLOR_ROW_HIGHLIGHTED, COLOR_ROW_SELECTED, COLOR_TARGET, JsonOutput, KeyValueTable,
-    L1BlocksTableParams, backlog_size_color, block_color, block_color_bright, build_gas_bar,
-    format_bytes, format_duration, format_gas, format_gwei, format_rate, format_unix_timestamp,
-    render_da_backlog_bar, render_gas_usage_bar, render_l1_blocks_table, target_usage_color,
-    time_diff_color, truncate_block_number,
+    L1BlocksTableParams, TimestampJson, backlog_size_color, block_color, block_color_bright,
+    build_gas_bar, format_bytes, format_duration, format_gas, format_gwei, format_rate,
+    format_unix_timestamp, render_da_backlog_bar, render_gas_usage_bar, render_l1_blocks_table,
+    target_usage_color, time_diff_color, truncate_block_number,
 };
 
 mod config;
